pyproject.toml
[tool.poetry]
name = "notebookmg"
version = "0.1.0"
description = "Opensource version of NotebookLm"
authors = ["TheM1N9 <manikanta9k7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
google-generativeai = "^0.8.3"
pypdf2 = "^3.0.1"
elevenlabs = "^1.50.3"
pydub = "^0.25.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
jinja2 = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
